Module invalid_mutations::bounds

Structs

ApplyCodeUnitBoundsContext

ApplyOutOfBoundsContext

CodeUnitBoundsMutation

Represents a single mutation onto a code unit to make it out of bounds.

OutOfBoundsMutation

Represents a single mutation to a CompiledModule to produce an out-of-bounds situation.

Enums

PointerKind

Represents the number of pointers that exist out from a node of a particular kind.

Statics

VALID_POINTER_SRCS